+++ libs/repo-api/test/auto-setup.test.ts
@@ -3,7 +3,7 @@ import { setup, Repo } from '@modtree/test-env'
 import { Api } from '../src'
 import '@modtree/test-env/jest'
 import { User, Degree, Graph } from '@modtree/entity'
-import { routes } from '../src/init'
+import { routes } from '../src/routes'
 import { postman } from '../src/postman'
 import fs from 'fs'
 import { join } from 'path'
